refactor: reorganize Author request classes and add type hints
